[Integration][Gitlab] Fix File Search Logging to only log when no files found (#1456)
# Description

**What -**  
This PR updates the asynchronous file search logic by introducing a flag
(`files_found`) to track whether any matching files are yielded during
iteration. If no files are found, it logs an appropriate message.

**Why -**  
Previously, the `async for...else` construct always executed the `else`
block, causing a log message to be printed even when files were
successfully found. The new approach ensures that the log message is
only triggered when no files match the search patterns.

**How -**  
1. A boolean flag `files_found` is initialized to `False` before
starting the asynchronous iteration.
2. As files are yielded by the asynchronous generator, the flag is set
to `True`.
3. After the iteration completes, the flag is checked. If no files were
found (`files_found` is still `False`), a log message is issued
indicating that no matching files were found.

## Type of change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] Non-breaking change (fix of existing functionality that will not
change current behavior)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] New Integration (non-breaking change which adds a new integration)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Documentation (added/updated documentation)

<h4> All tests should be run against the port production
environment(using a testing org). </h4>

### Core testing checklist

- [ ] Integration able to create all default resources from scratch
- [ ] Resync finishes successfully
- [ ] Resync able to create entities
- [ ] Resync able to update entities
- [ ] Resync able to detect and delete entities
- [ ] Scheduled resync able to abort existing resync and start a new one
- [ ] Tested with at least 2 integrations from scratch
- [ ] Tested with Kafka and Polling event listeners
- [ ] Tested deletion of entities that don't pass the selector

### Integration testing checklist

- [ ] Integration able to create all default resources from scratch
- [ ] Resync able to create entities
- [ ] Resync able to update entities
- [ ] Resync able to detect and delete entities
- [ ] Resync finishes successfully
- [ ] If new resource kind is added or updated in the integration, add
example raw data, mapping and expected result to the `examples` folder
in the integration directory.
- [ ] If resource kind is updated, run the integration with the example
data and check if the expected result is achieved
- [ ] If new resource kind is added or updated, validate that
live-events for that resource are working as expected
- [ ] Docs PR link [here](#)

### Preflight checklist

- [ ] Handled rate limiting
- [ ] Handled pagination
- [ ] Implemented the code in async
- [ ] Support Multi account

## Screenshots

Include screenshots from your environment showing how the resources of
the integration will look.

## API Documentation

Provide links to the API documentation used for this integration.