[Integration][AWS] Resolve Cloud Control Pagination Issue (#1453)
# Description

**What**  
- This PR fixes an issue in the AWS integration where the API returns
only 1000 resources in a single batch for the SQS Kind, despite the
MaxResults parameter being set, and no NextToken is provided to paginate
through all available resources.
- Some resources have uncontrollable numbers of resources retrieved per
paginated request.

**Why**  
The current behavior leads to incomplete ingestion (e.g., tests made it
clear that, only 1000 out of 1500 SQS queues available are processed)
since the MaxResults parameter is ignored and subsequent pages cannot be
retrieved. This bug, observed on the AWS SQS kind and potentially
affecting other resources using the cloudcontrol API, needs to be
resolved to ensure that all resources are ingested as expected.

**How**  
- Introduced a custom pagination utility `AsyncPaginator` to better
manage paginated API requests.
- Modified the integration to detach SQS from the cloudcontrol API,
addressing the missing `NextToken` issue that previously prevented
further pagination.
- The new utility ensures that regardless of the response from the API,
the given Batch Size is respected during pagination.

## Type of change

- [x] Bug fix (non-breaking change which fixes an issue)

<h4> All tests should be run against the port production environment
(using a testing org). </h4>

### Core testing checklist

- [ ] Integration able to create all default resources from scratch
- [ ] Resync finishes successfully
- [ ] Resync able to create entities
- [ ] Resync able to update entities
- [ ] Resync able to detect and delete entities
- [ ] Scheduled resync able to abort existing resync and start a new one
- [ ] Tested with at least 2 integrations from scratch
- [ ] Tested with Kafka and Polling event listeners
- [ ] Tested deletion of entities that don't pass the selector

### Integration testing checklist

- [ ] Integration able to create all default resources from scratch
- [ ] Resync able to create entities
- [ ] Resync able to update entities
- [ ] Resync able to detect and delete entities
- [ ] Resync finishes successfully
- [ ] If new resource kind is added or updated in the integration, add
example raw data, mapping and expected result to the `examples` folder
in the integration directory.
- [ ] If resource kind is updated, run the integration with the example
data and check if the expected result is achieved
- [ ] If new resource kind is added or updated, validate that
live-events for that resource are working as expected
- [ ] Docs PR link [here](#)

### Preflight checklist

- [ ] Handled rate limiting
- [ ] Handled pagination
- [ ] Implemented the code in async
- [ ] Support Multi account

## Screenshots

Include screenshots from your environment showing how the resources of
the integration will look.

## API Documentation

Provide links to the API documentation used for this integration.

---------

Co-authored-by: Tom Tankilevitch <59158507+Tankilevitch@users.noreply.github.com>