[Integration][Github] - Implement support for Github App (#1671)
# Description
This add support for Github apps to our Ocean integration for Github.

What - Settings and configurations to authenticate as a Github app.

Why - We want to make the experience of using the Ocean integration as
similar as the existing app, this means giving the integration the
capability of being installed with our existing app flow. This is the
first step to that goal. At the moment this works more like the
"self-hosted app" process.

How - This works by subclassing the existing GithubRestClient and adding
App-specific functionalities on top. The existing abstraction means the
support can be added with minimal changes.

## Type of change

Please leave one option from the following and delete the rest:

- [ x ] New feature (non-breaking change which adds functionality)

<h4> All tests should be run against the port production
environment(using a testing org). </h4>

### Core testing checklist

- [ ] Integration able to create all default resources from scratch
- [ ] Resync finishes successfully
- [ ] Resync able to create entities
- [ ] Resync able to update entities
- [ ] Resync able to detect and delete entities
- [ ] Scheduled resync able to abort existing resync and start a new one
- [ ] Tested with at least 2 integrations from scratch
- [ ] Tested with Kafka and Polling event listeners
- [ ] Tested deletion of entities that don't pass the selector


### Integration testing checklist

- [ ] Integration able to create all default resources from scratch
- [ ] Resync able to create entities
- [ ] Resync able to update entities
- [ ] Resync able to detect and delete entities
- [ ] Resync finishes successfully
- [ ] If new resource kind is added or updated in the integration, add
example raw data, mapping and expected result to the `examples` folder
in the integration directory.
- [ ] If resource kind is updated, run the integration with the example
data and check if the expected result is achieved
- [ ] If new resource kind is added or updated, validate that
live-events for that resource are working as expected
- [ ] Docs PR link [here](#)

### Preflight checklist

- [ ] Handled rate limiting
- [ ] Handled pagination
- [ ] Implemented the code in async
- [ ] Support Multi account

## Screenshots

Include screenshots from your environment showing how the resources of
the integration will look.

## API Documentation

Provide links to the API documentation used for this integration.

---------

Co-authored-by: Michael Kofi Armah <mikeyarmah@gmail.com>
Co-authored-by: emekanwaoma <chukwuemeka.nwaoma@port.io>