[Core] Split search execution to sync and async
Update the core to handle jq calculations without a loop, the search functions that need to be executed by the overwriten function will fail the compile. Thaw way we can execute those mappings in the event loop since we can assume that they require the overwritten function.
There is a filter that will stop the broken mappings.
Later we merge the results returned by gather, they are ordered by the queuing order so we can safely assume that we can merge them.